Disable "setDataFasterThanImageUpload" in D3D11.
For some reason using setData routine causes huge memory issues
on some hardware / software combinations. In a reproducible way it
looked like that:

XPS, Integrated Intel graphics card with shared with the process VRAM.

When in Telegram Desktop media viewer some QWidget was animating,
the memory usage was growing very fast, consuming up to 15 GB in
around 10 second animation. While trying to debug this I narrowed
it down to UpdateSubresource() calls, which looked like each call
was using new memory without freeing the previous, although the calls
were made for the same texture on the same context over and over.

The taken memory was eventually freed by the driver without any
actions from the application side, in a different timeframe, but
within a minute guaranteed. So there was no leak on the application
side, just some strange sort of clearing deferring in the driver.

This was even more serious for 32bit build of Telegram Desktop,
because in this case the memory usage was quickly growing to > 3 GB
and then the device was lost and the media viewer stopped updating.
The application continued working if no large allocation
(leading to OOM) was triggered, then the memory was cleared within a
minute and the app could continue working as it should, but with a
broken media viewer.

In case "Image11::loadData" routine is used, the memory usage doesn't
grow that way, instead it works correctly, clearing instantly.